+++ multi_thread/autoload.php
@@ -1,0 +1,7 @@
+<?php
+
+require_once __DIR__ . '/TaskResult.php';
+require_once __DIR__ . '/TaskRunnerInterface.php';
+require_once __DIR__ . '/TaskRunner.php';
+require_once __DIR__ . '/CurlMultiTaskRunner.php';
+require_once __DIR__ . '/ProcessTaskRunner.php';